Refactor redundant chunks of code
