+++ deeplabcut/gui/multiple_individuals_labeling_toolbox.py
@@ -15,7 +15,6 @@ import os.path
 from pathlib import Path
 
 import cv2
-import matplotlib
 import matplotlib.colors as mcolors
 import matplotlib.patches as patches
 import matplotlib.pyplot as plt
@@ -23,38 +22,20 @@ import numpy as np
 import pandas as pd
 import wx
 import wx.lib.scrolledpanel as SP
-from matplotlib.backends.backend_wxagg import FigureCanvasWxAgg as FigureCanvas
 from matplotlib.backends.backend_wxagg import (
     NavigationToolbar2WxAgg as NavigationToolbar,
 )
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 from deeplabcut.gui import auxfun_drag
+from deeplabcut.gui.widgets import BasePanel, WidgetPanel, BaseFrame
 from deeplabcut.utils import auxiliaryfunctions, auxfun_multianimal
 
 
 # ###########################################################################
 # Class for GUI MainFrame
 # ###########################################################################
-class ImagePanel(wx.Panel):
-    def __init__(self, parent, config, gui_size, **kwargs):
-        h = gui_size[0] / 2
-        w = gui_size[1] / 3
-        wx.Panel.__init__(self, parent, -1, style=wx.SUNKEN_BORDER, size=(h, w))
-
-        self.figure = matplotlib.figure.Figure()
-        self.axes = self.figure.add_subplot(1, 1, 1)
-        self.canvas = FigureCanvas(self, -1, self.figure)
-        self.orig_xlim = None
-        self.orig_ylim = None
-        self.sizer = wx.BoxSizer(wx.VERTICAL)
-        self.sizer.Add(self.canvas, 1, wx.LEFT | wx.TOP | wx.GROW)
-        self.SetSizer(self.sizer)
-        self.Fit()
-
-    def getfigure(self):
-        return self.figure
-
+class ImagePanel(BasePanel):
     def drawplot(self, img, img_name, itr, index, bodyparts, cmap, keep_view=False):
         xlim = self.axes.get_xlim()
         ylim = self.axes.get_ylim()
@@ -91,10 +72,6 @@ class ImagePanel(wx.Panel):
         self.toolbar = NavigationToolbar(self.canvas)
         return (self.figure, self.axes, self.canvas, self.toolbar)
 
-    def resetView(self):
-        self.axes.set_xlim(self.orig_xlim)
-        self.axes.set_ylim(self.orig_ylim)
-
     def getColorIndices(self, img, bodyparts):
         """
         Returns the colormaps ticks and . The order of ticks labels is reversed.
@@ -103,11 +80,6 @@ class ImagePanel(wx.Panel):
         norm = mcolors.Normalize(vmin=0, vmax=np.max(im))
         ticks = np.linspace(0, np.max(im), len(bodyparts))[::-1]
         return norm, ticks
-
-
-class WidgetPanel(wx.Panel):
-    def __init__(self, parent):
-        wx.Panel.__init__(self, parent, -1, style=wx.SUNKEN_BORDER)
 
 
 class ScrollPanel(SP.ScrolledPanel):
@@ -171,40 +143,17 @@ class ScrollPanel(SP.ScrolledPanel):
         self.choiceBox.Clear(True)
 
 
-class MainFrame(wx.Frame):
-    """Contains the main GUI and button boxes"""
-
+class MainFrame(BaseFrame):
     def __init__(self, parent, config):
-        # Settting the GUI size and panels design
-        displays = (
-            wx.Display(i) for i in range(wx.Display.GetCount())
-        )  # Gets the number of displays
-        screenSizes = [
-            display.GetGeometry().GetSize() for display in displays
-        ]  # Gets the size of each display
-        index = 0  # For display 1.
-        screenWidth = screenSizes[index][0]
-        screenHeight = screenSizes[index][1]
-        self.gui_size = (screenWidth * 0.7, screenHeight * 0.85)
-
-        wx.Frame.__init__(
-            self,
-            parent,
-            id=wx.ID_ANY,
-            title="DeepLabCut2.0 - Multiple Individuals Labeling ToolBox",
-            size=wx.Size(self.gui_size),
-            pos=wx.DefaultPosition,
-            style=wx.RESIZE_BORDER | wx.DEFAULT_FRAME_STYLE | wx.TAB_TRAVERSAL,
+        super(MainFrame, self).__init__(
+            "DeepLabCut2.0 - Multiple Individuals Labeling ToolBox", parent,
         )
-        self.statusbar = self.CreateStatusBar()
+
         self.statusbar.SetStatusText(
             "Looking for a folder to start labeling. Click 'Load frames' to begin."
         )
         self.Bind(wx.EVT_CHAR_HOOK, self.OnKeyPressed)
 
-        self.SetSizeHints(
-            wx.Size(self.gui_size)
-        )  #  This sets the minimum size of the GUI. It can scale now!
         ###################################################################################################################################################
 
         # Spliting the frame into top and bottom panels. Bottom panels contains the widgets. The top panel is for showing images and plotting!
@@ -346,10 +295,6 @@ class MainFrame(wx.Frame):
         elif event.ControlDown() and event.GetKeyCode() == 67:
             self.duplicate_labels()
 
-    @staticmethod
-    def calc_distance(x1, y1, x2, y2):
-        return np.sqrt((x1 - x2) ** 2 + (y1 - y2) ** 2)
-
     def duplicate_labels(self):
         if self.iter >= 1:
             curr_individual = self.individualrdb.GetStringSelection()
@@ -469,56 +414,10 @@ class MainFrame(wx.Frame):
         )
         self.statusbar.SetStatusText("Help")
 
-    def homeButton(self, event):
-        self.image_panel.resetView()
-        self.figure.canvas.draw()
-        MainFrame.updateZoomPan(self)
-        self.zoom.SetValue(False)
-        self.pan.SetValue(False)
-        self.statusbar.SetStatusText("")
-
-    def panButton(self, event):
-        if self.pan.GetValue():
-            self.toolbar.pan()
-            self.statusbar.SetStatusText("Pan On")
-            self.zoom.SetValue(False)
-        else:
-            self.toolbar.pan()
-            self.statusbar.SetStatusText("Pan Off")
-
-    def zoomButton(self, event):
-        if self.zoom.GetValue():
-            # Save pre-zoom xlim and ylim values
-            self.prezoom_xlim = self.axes.get_xlim()
-            self.prezoom_ylim = self.axes.get_ylim()
-            self.toolbar.zoom()
-            self.statusbar.SetStatusText("Zoom On")
-            self.pan.SetValue(False)
-        else:
-            self.toolbar.zoom()
-            self.statusbar.SetStatusText("Zoom Off")
-
-    def onZoom(self, ax):
-        # See if axis limits have actually changed
-        curr_xlim = self.axes.get_xlim()
-        curr_ylim = self.axes.get_ylim()
-        if self.zoom.GetValue() and not (
-            self.prezoom_xlim[0] == curr_xlim[0]
-            and self.prezoom_xlim[1] == curr_xlim[1]
-            and self.prezoom_ylim[0] == curr_ylim[0]
-            and self.prezoom_ylim[1] == curr_ylim[1]
-        ):
-            self.updateZoomPan()
-            self.statusbar.SetStatusText("Zoom Off")
-
     def onButtonRelease(self, event):
         if self.pan.GetValue():
             self.updateZoomPan()
             self.statusbar.SetStatusText("Pan Off")
-
-    def lockChecked(self, event):
-        self.cb = event.GetEventObject()
-        self.view_locked = self.cb.GetValue()
 
     def onClick(self, event):
         """
@@ -1270,15 +1169,6 @@ class MainFrame(wx.Frame):
             self.cidClick = self.canvas.mpl_connect("button_press_event", self.onClick)
         else:
             self.change_marker_size.Enable(False)
-
-    def updateZoomPan(self):
-        # Checks if zoom/pan button is ON
-        if self.pan.GetValue():
-            self.toolbar.pan()
-            self.pan.SetValue(False)
-        if self.zoom.GetValue():
-            self.toolbar.zoom()
-            self.zoom.SetValue(False)
 
 
 def show(config):
