+++ deeplabcut/gui/multiple_individuals_refinement_toolbox.py
@@ -16,7 +16,6 @@ from pathlib import Path
 
 # from skimage import io
 import PIL
-import matplotlib
 import matplotlib.colors as mcolors
 import matplotlib.patches as patches
 import matplotlib.pyplot as plt
@@ -24,7 +23,6 @@ import numpy as np
 import pandas as pd
 import wx
 import wx.lib.scrolledpanel as SP
-from matplotlib.backends.backend_wxagg import FigureCanvasWxAgg as FigureCanvas
 from matplotlib.backends.backend_wxagg import (
     NavigationToolbar2WxAgg as NavigationToolbar,
 )
@@ -32,31 +30,14 @@ from mpl_toolkits.axes_grid1 import make_axes_locatable
 from skimage import io
 
 from deeplabcut.gui import auxfun_drag
+from deeplabcut.gui.widgets import BasePanel, WidgetPanel, BaseFrame
 from deeplabcut.utils import auxiliaryfunctions, visualization
 
 
 # ###########################################################################
 # Class for GUI MainFrame
 # ###########################################################################
-class ImagePanel(wx.Panel):
-    def __init__(self, parent, config, gui_size, **kwargs):
-        h = gui_size[0] / 2
-        w = gui_size[1] / 3
-        wx.Panel.__init__(self, parent, -1, style=wx.SUNKEN_BORDER, size=(h, w))
-
-        self.figure = matplotlib.figure.Figure()
-        self.axes = self.figure.add_subplot(1, 1, 1)
-        self.canvas = FigureCanvas(self, -1, self.figure)
-        self.orig_xlim = None
-        self.orig_ylim = None
-        self.sizer = wx.BoxSizer(wx.VERTICAL)
-        self.sizer.Add(self.canvas, 1, wx.LEFT | wx.TOP | wx.GROW)
-        self.SetSizer(self.sizer)
-        self.Fit()
-
-    def getfigure(self):
-        return self.figure
-
+class ImagePanel(BasePanel):
     def drawplot(
         self, img, img_name, itr, index, threshold, cmap, preview, keep_view=False
     ):
@@ -97,10 +78,6 @@ class ImagePanel(wx.Panel):
         self.toolbar = NavigationToolbar(self.canvas)
         return (self.figure, self.axes, self.canvas, self.toolbar, self.ax)
 
-    def resetView(self):
-        self.axes.set_xlim(self.orig_xlim)
-        self.axes.set_ylim(self.orig_ylim)
-
     def getColorIndices(self, img, bodyparts):
         """
         Returns the colormaps ticks and . The order of ticks labels is reversed.
@@ -109,11 +86,6 @@ class ImagePanel(wx.Panel):
         norm = mcolors.Normalize(vmin=np.min(im), vmax=np.max(im))
         ticks = np.linspace(np.min(im), np.max(im), len(bodyparts))[::-1]
         return norm, ticks
-
-
-class WidgetPanel(wx.Panel):
-    def __init__(self, parent):
-        wx.Panel.__init__(self, parent, -1, style=wx.SUNKEN_BORDER)
 
 
 class ScrollPanel(SP.ScrolledPanel):
@@ -162,38 +134,13 @@ class ScrollPanel(SP.ScrolledPanel):
         self.choiceBox.Clear(True)
 
 
-class MainFrame(wx.Frame):
-    """Contains the main GUI and button boxes"""
-
+class MainFrame(BaseFrame):
     def __init__(self, parent, config):
-        # Settting the GUI size and panels design
-        displays = (
-            wx.Display(i) for i in range(wx.Display.GetCount())
-        )  # Gets the number of displays
-        screenSizes = [
-            display.GetGeometry().GetSize() for display in displays
-        ]  # Gets the size of each display
-        index = 0  # For display 1.
-        screenWidth = screenSizes[index][0]
-        screenHeight = screenSizes[index][1]
-        self.gui_size = (screenWidth * 0.7, screenHeight * 0.85)
-
-        wx.Frame.__init__(
-            self,
-            parent,
-            id=wx.ID_ANY,
-            title="DeepLabCut - Refinement ToolBox",
-            size=wx.Size(self.gui_size),
-            pos=wx.DefaultPosition,
-            style=wx.RESIZE_BORDER | wx.DEFAULT_FRAME_STYLE | wx.TAB_TRAVERSAL,
+        super(MainFrame, self).__init__(
+            "DeepLabCut - Refinement ToolBox", parent,
         )
-        self.statusbar = self.CreateStatusBar()
-        self.statusbar.SetStatusText("")
         self.Bind(wx.EVT_CHAR_HOOK, self.OnKeyPressed)
 
-        self.SetSizeHints(
-            wx.Size(self.gui_size)
-        )  #  This sets the minimum size of the GUI. It can scale now!
         ###################################################################################################################################################
 
         # Spliting the frame into top and bottom panels. Bottom panels contains the widgets. The top panel is for showing images and plotting!
@@ -347,55 +294,6 @@ class MainFrame(wx.Frame):
             if msg == 2:
                 closest_dp.delete_data()
 
-    @staticmethod
-    def calc_distance(x1, y1, x2, y2):
-        return np.sqrt((x1 - x2) ** 2 + (y1 - y2) ** 2)
-
-    def closewindow(self, event):
-        self.Destroy()
-
-    def homeButton(self, event):
-        self.image_panel.resetView()
-        self.figure.canvas.draw()
-        MainFrame.updateZoomPan(self)
-        self.zoom.SetValue(False)
-        self.pan.SetValue(False)
-        self.statusbar.SetStatusText("")
-
-    def panButton(self, event):
-        if self.pan.GetValue():
-            self.toolbar.pan()
-            self.statusbar.SetStatusText("Pan On")
-            self.zoom.SetValue(False)
-        else:
-            self.toolbar.pan()
-            self.statusbar.SetStatusText("Pan Off")
-
-    def zoomButton(self, event):
-        if self.zoom.GetValue():
-            # Save pre-zoom xlim and ylim values
-            self.prezoom_xlim = self.axes.get_xlim()
-            self.prezoom_ylim = self.axes.get_ylim()
-            self.toolbar.zoom()
-            self.statusbar.SetStatusText("Zoom On")
-            self.pan.SetValue(False)
-        else:
-            self.toolbar.zoom()
-            self.statusbar.SetStatusText("Zoom Off")
-
-    def onZoom(self, ax):
-        # See if axis limits have actually changed
-        curr_xlim = self.axes.get_xlim()
-        curr_ylim = self.axes.get_ylim()
-        if self.zoom.GetValue() and not (
-            self.prezoom_xlim[0] == curr_xlim[0]
-            and self.prezoom_xlim[1] == curr_xlim[1]
-            and self.prezoom_ylim[0] == curr_ylim[0]
-            and self.prezoom_ylim[1] == curr_ylim[1]
-        ):
-            self.updateZoomPan()
-            self.statusbar.SetStatusText("Zoom Off")
-
     def activateSlider(self, event):
         """
         Activates the slider to increase the markersize
@@ -440,10 +338,6 @@ class MainFrame(wx.Frame):
         self.axes.callbacks.connect("ylim_changed", self.onZoom)
 
         MainFrame.plot(self, self.img)
-
-    def lockChecked(self, event):
-        self.cb = event.GetEventObject()
-        self.view_locked = self.cb.GetValue()
 
     def browseDir(self, event):
         """
@@ -1261,15 +1155,6 @@ class MainFrame(wx.Frame):
                         self.drs.append(self.dr)
                         self.updatedCoords.append(self.dr.coords)
         self.figure.canvas.draw()
-
-    def updateZoomPan(self):
-        # Checks if zoom/pan button is ON
-        if self.pan.GetValue():
-            self.toolbar.pan()
-            self.pan.SetValue(False)
-        if self.zoom.GetValue():
-            self.toolbar.zoom()
-            self.zoom.SetValue(False)
 
 
 def show(config):
