+++ deeplabcut/gui/select_crop_parameters.py
@@ -10,79 +10,23 @@ Licensed under GNU Lesser General Public License v3.0
 
 import argparse
 
-import matplotlib
 import wx
-from matplotlib.backends.backend_wxagg import FigureCanvasWxAgg as FigureCanvas
 from matplotlib.figure import Figure
 from matplotlib.widgets import RectangleSelector
+from deeplabcut.gui.widgets import BasePanel, WidgetPanel, BaseFrame
 
 
-# ###########################################################################
-# Class for GUI MainFrame
-# ###########################################################################
-class ImagePanel(wx.Panel):
-    def __init__(self, parent, config, gui_size, **kwargs):
-        h = gui_size[0] / 2
-        w = gui_size[1] / 3
-        wx.Panel.__init__(self, parent, -1, style=wx.SUNKEN_BORDER, size=(h, w))
-
-        self.figure = matplotlib.figure.Figure()
-        self.axes = self.figure.add_subplot(1, 1, 1)
-        self.canvas = FigureCanvas(self, -1, self.figure)
-        self.sizer = wx.BoxSizer(wx.VERTICAL)
-        self.sizer.Add(self.canvas, 1, wx.LEFT | wx.TOP | wx.GROW)
-        self.SetSizer(self.sizer)
-        self.Fit()
-
-    def getfigure(self):
-        """
-        Returns the figure, axes and canvas
-        """
-        return (self.figure, self.axes, self.canvas)
-
-
-class WidgetPanel(wx.Panel):
-    def __init__(self, parent):
-        wx.Panel.__init__(self, parent, -1, style=wx.SUNKEN_BORDER)
-
-
-class MainFrame(wx.Frame):
-    """Contains the main GUI and button boxes"""
-
+class MainFrame(BaseFrame):
     def __init__(self, parent, config, image):
-        # Settting the GUI size and panels design
-        displays = (
-            wx.Display(i) for i in range(wx.Display.GetCount())
-        )  # Gets the number of displays
-        screenSizes = [
-            display.GetGeometry().GetSize() for display in displays
-        ]  # Gets the size of each display
-        index = 0  # For display 1.
-        screenWidth = screenSizes[index][0]
-        screenHeight = screenSizes[index][1]
-        self.gui_size = (screenWidth * 0.7, screenHeight * 0.85)
-
-        wx.Frame.__init__(
-            self,
-            parent,
-            id=wx.ID_ANY,
-            title="DeepLabCut2.0 - Select Crop Parameters",
-            size=wx.Size(self.gui_size),
-            pos=wx.DefaultPosition,
-            style=wx.RESIZE_BORDER | wx.DEFAULT_FRAME_STYLE | wx.TAB_TRAVERSAL,
+        super(MainFrame, self).__init__(
+            "DeepLabCut2.0 - Select Crop Parameters", parent,
         )
-        self.statusbar = self.CreateStatusBar()
-        self.statusbar.SetStatusText("")
-
-        self.SetSizeHints(
-            wx.Size(self.gui_size)
-        )  #  This sets the minimum size of the GUI. It can scale now!
 
         ###################################################################################################################################################
         # Spliting the frame into top and bottom panels. Bottom panels contains the widgets. The top panel is for showing images and plotting!
         topSplitter = wx.SplitterWindow(self)
 
-        self.image_panel = ImagePanel(topSplitter, config, self.gui_size)
+        self.image_panel = BasePanel(topSplitter, config, self.gui_size)
         self.widget_panel = WidgetPanel(topSplitter)
 
         topSplitter.SplitHorizontally(
